Added a DBAL DataSource accepting a connection, updated Doctrine DataSource to use it
